+++ components/config/src/highlighting.rs
@@ -16,29 +16,40 @@ lazy_static! {
         from_binary(include_bytes!("../../../sublime/themes/all.themedump"));
 }
 
+pub enum HighlightSource {
+    Theme,
+    Extra,
+    Plain,
+    NotFound,
+}
+
 /// Returns the highlighter and whether it was found in the extra or not
-pub fn get_highlighter(language: Option<&str>, config: &Config) -> (HighlightLines<'static>, bool) {
+pub fn get_highlighter(
+    language: Option<&str>,
+    config: &Config,
+) -> (HighlightLines<'static>, HighlightSource) {
     let theme = &THEME_SET.themes[config.highlight_theme()];
-    let mut in_extra = false;
 
     if let Some(ref lang) = language {
-        let syntax = if let Some(ref extra) = config.markdown.extra_syntax_set {
-            let s = extra.find_syntax_by_token(lang);
-            if s.is_some() {
-                in_extra = true;
+        if let Some(ref extra_syntaxes) = config.markdown.extra_syntax_set {
+            if let Some(syntax) = extra_syntaxes.find_syntax_by_token(lang) {
+                return (HighlightLines::new(syntax, theme), HighlightSource::Extra);
             }
-            s
-        } else {
-            // The JS syntax hangs a lot... the TS syntax is probably better anyway.
-            // https://github.com/getzola/zola/issues/1241
-            // https://github.com/getzola/zola/issues/1211
-            // https://github.com/getzola/zola/issues/1174
-            let hacked_lang = if *lang == "js" || *lang == "javascript" { "ts" } else { lang };
-            SYNTAX_SET.find_syntax_by_token(hacked_lang)
         }
-        .unwrap_or_else(|| SYNTAX_SET.find_syntax_plain_text());
-        (HighlightLines::new(syntax, theme), in_extra)
+        // The JS syntax hangs a lot... the TS syntax is probably better anyway.
+        // https://github.com/getzola/zola/issues/1241
+        // https://github.com/getzola/zola/issues/1211
+        // https://github.com/getzola/zola/issues/1174
+        let hacked_lang = if *lang == "js" || *lang == "javascript" { "ts" } else { lang };
+        if let Some(syntax) = SYNTAX_SET.find_syntax_by_token(hacked_lang) {
+            (HighlightLines::new(syntax, theme), HighlightSource::Theme)
+        } else {
+            (
+                HighlightLines::new(SYNTAX_SET.find_syntax_plain_text(), theme),
+                HighlightSource::NotFound,
+            )
+        }
     } else {
-        (HighlightLines::new(SYNTAX_SET.find_syntax_plain_text(), theme), false)
+        (HighlightLines::new(SYNTAX_SET.find_syntax_plain_text(), theme), HighlightSource::Plain)
     }
 }