+++ components/config/src/lib.rs
@@ -4,18 +4,12 @@ mod theme;
 pub use crate::config::{
     languages::Language, link_checker::LinkChecker, slugify::Slugify, taxonomies::Taxonomy, Config,
 };
+use errors::Result;
 
 use std::path::Path;
 
 /// Get and parse the config.
 /// If it doesn't succeed, exit
-pub fn get_config(filename: &Path) -> Config {
-    match Config::from_file(filename) {
-        Ok(c) => c,
-        Err(e) => {
-            println!("Failed to load {}", filename.display());
-            println!("Error: {}", e);
-            ::std::process::exit(1);
-        }
-    }
+pub fn get_config(filename: &Path) -> Result<Config> {
+    Config::from_file(filename)
 }