+++ components/library/src/content/page.rs
@@ -62,6 +62,10 @@ pub struct Page {
     pub earlier: Option<DefaultKey>,
     /// The later page, for pages sorted by date
     pub later: Option<DefaultKey>,
+    /// The previous page, for pages sorted by title
+    pub title_prev: Option<DefaultKey>,
+    /// The next page, for pages sorted by title
+    pub title_next: Option<DefaultKey>,
     /// The lighter page, for pages sorted by weight
     pub lighter: Option<DefaultKey>,
     /// The heavier page, for pages sorted by weight
@@ -486,7 +490,7 @@ Hello world"#;
         let mut config = Config::default();
         config.slugify.paths = SlugifyStrategy::On;
         let res =
-            Page::parse(Path::new(" file with space.md"), "+++\n+++", &config, &PathBuf::new());
+            Page::parse(Path::new(" file with space.md"), "+++\n+++\n", &config, &PathBuf::new());
         assert!(res.is_ok());
         let page = res.unwrap();
         assert_eq!(page.slug, "file-with-space");
@@ -497,7 +501,7 @@ Hello world"#;
     fn can_make_path_from_utf8_filename() {
         let mut config = Config::default();
         config.slugify.paths = SlugifyStrategy::Safe;
-        let res = Page::parse(Path::new("日本.md"), "+++\n++++", &config, &PathBuf::new());
+        let res = Page::parse(Path::new("日本.md"), "+++\n+++\n", &config, &PathBuf::new());
         assert!(res.is_ok());
         let page = res.unwrap();
         assert_eq!(page.slug, "日本");